rename dossier folder, refs 347
