+++ src/app/shared/components/student-dossier/dossier-overview/dossier-overview.component.scss
