+++ src/app/shared/components/student-dossier/student-profile-absences/student-profile-absences.component.scss
