+++ src/app/shared/components/student-dossier/student-profile-apprenticeship-company/student-profile-apprenticeship-company.component.spec.ts
