+++ src/app/shared/components/student-dossier/student-profile-entry-header/student-profile-entry-header.component.scss
